rename braid_bench / chan_bench & improve formatting
